Separated the tabnav creation from the tabnav rendering, so that you can create a tabnav without immediately rendering it.
This allows you, for example, to query the tabnav to see which tab is
highlighted and to show different content (such as a different sub-tabnav)
depending on the active tab. This keeps things DRY.

No longer assuming that each tabnav will will be defined in a partial in
app/views/widgets. This allows you the flexibility, for example, to
define all your tabnavs in a single file, or to define a tabnav that is
only used in a single view at the top of that view.

Refactored highlighted? to use a new, more general method, matches_path?.

Added highlights_not_on method to
Widgets::Highlightable::InstanceMethods to let you specify paths for
which a tab should *never* be highlighted for.

Added a 'tabnav' class to the outer div and the ul to allow stylesheets
and jQuery selectors to easily select them.

Added some tests.